Added setup, restructured built-in checks
Added setup.py
Restructured package for sanity
Moved RobotsTest out of corechecks because it is lame.
Moved djangochecks out of the package (will be put into a separate repo)
Added WebTouch and Http200 to corechecks (WebTouch simply visits an URL
with a get, and follows a redirect if present, Http200 visits a URL and
passes if it returns 200, but doesn't follow redirects.